RandomMission component code splitting
