+++ src/app/profile/[id]/setting/page.tsx
@@ -17,15 +17,14 @@ const ModifyingProgile = () => {
   const [number, setNumber] = useState("");
   const [birthday, setBirthday] = useState("");
 
-  const params = useParams();
-  const searchId = decodeURIComponent(`${params.id}`);
+  const searchId = useParams().id
+  console.log("searchId==>",searchId)
 
   const getUser = async () => {
     const { data } = await supabase
       .from("user")
       .select()
-      .eq("nickname", searchId);
-
+      .eq("uid", searchId);
     setUser(data);
     setUserInfo(data);
 
@@ -45,7 +44,7 @@ const ModifyingProgile = () => {
     const { error: updateDataError } = await supabase
       .from("user")
       .update({ nickname: name, email, number, birthday })
-      .eq("nickname", searchId);
+      .eq("uid", searchId);
 
     alert("수정이 완료되었습니다.");
   };
